Separate installer tests and update tool tests
