+++ test-scripts/ballerina-test-automation/build.gradle
@@ -27,28 +27,8 @@ version '1.0-SNAPSHOT'
 
 sourceCompatibility = 1.8
 
-dependencies {
-    implementation 'org.slf4j:slf4j-api:1.7.26'
-    implementation 'info.picocli:picocli:4.0.1'
-    implementation 'org.testng:testng:6.14.3'
-    testCompile 'org.testng:testng'
-}
-
-sourceCompatibility = '1.8'
-
 tasks.withType(JavaCompile) {
     options.encoding = 'UTF-8'
-}
-
-test {
-    useTestNG()
-    options.suites('src/test/resources/testng.xml')
-    systemProperty 'BALLERINA_VERSION', project.properties['ballerinaVersion']
-    systemProperty 'SPEC_VERSION', project.properties['specVersion']
-    systemProperty 'TOOL_VERSION', project.properties['toolVersion']
-    systemProperty 'LATEST_TOOL_VERSION', project.properties['latestToolVersion']
-    systemProperty 'LATEST_PATCH_VERSION', project.properties['latestPatchVersion']
-    systemProperty 'LATEST_PATCH_SPEC_VERSION', project.properties['latestPatchSpecVersion']
 }
 
 repositories {
