+++ test-scripts/ballerina-test-automation/installer-test/src/test/java/io/ballerina/installer/test/CentralTest.java
@@ -14,21 +14,16 @@
  * limitations under the License.
  */
 
-package io.ballerina.test;
+package io.ballerina.installer.test;
 
+import io.ballerina.test.Executor;
+import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 
-public class UpdateDistTest {
+public class CentralTest {
     String version = System.getProperty("BALLERINA_VERSION");
-    String specVersion = System.getProperty("SPEC_VERSION");
-    String toolVersion = System.getProperty("TOOL_VERSION");
-    String latestToolVersion = System.getProperty("LATEST_TOOL_VERSION");
-
-    String previousVersion = "1.2.0";
-    String previousSpecVersion = "2020R1";
-    String previousVersionsLatestPatch = System.getProperty("LATEST_PATCH_VERSION");
 
     @DataProvider(name = "getExecutors")
     public Object[][] dataProviderMethod() {
@@ -38,13 +33,12 @@ public class UpdateDistTest {
     }
 
     @Test(dataProvider = "getExecutors")
-    public void testDistCommands(Executor executor) {
+    public void testPull(Executor executor) {
         executor.transferArtifacts();
         executor.install();
-
-        TestUtils.testDistCommands(executor, version, specVersion, toolVersion, previousVersion, previousSpecVersion,
-                previousVersionsLatestPatch, latestToolVersion);
-
+        //Checks part as output varies depending on the network speed
+        Assert.assertTrue(executor.executeCommand("ballerina pull ballerinax/sdfc", false)
+                .contains("pulled from central successfully"));
         executor.uninstall();
         executor.cleanArtifacts();
     }
