+++ test-scripts/ballerina-test-automation/installer-test/src/test/java/io/ballerina/installer/test/InstallerTest.java
@@ -14,48 +14,31 @@
  * limitations under the License.
  */
 
-package io.ballerina.test;
+package io.ballerina.installer.test;
 
-import org.testng.Assert;
+import io.ballerina.test.Executor;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 
-public class UpdateToolTest {
+public class InstallerTest {
     String version = System.getProperty("BALLERINA_VERSION");
     String specVersion = System.getProperty("SPEC_VERSION");
     String toolVersion = System.getProperty("TOOL_VERSION");
-    String latestToolVersion = System.getProperty("LATEST_TOOL_VERSION");
-
-    String previousVersion = "1.2.0";
-    String previousSpecVersion = "2020R1";
-    String previousVersionsLatestPatch = System.getProperty("LATEST_PATCH_VERSION");
-    String previousToolVersion = "0.8.5";
 
     @DataProvider(name = "getExecutors")
     public Object[][] dataProviderMethod() {
         Executor[][] result = new Executor[1][1];
-        result[0][0] = TestUtils.getExecutor(previousVersion);
+        result[0][0] = TestUtils.getExecutor(version);
         return result;
     }
 
     @Test(dataProvider = "getExecutors")
-    public void testUpdateTool(Executor executor) {
+    public void testSmoke(Executor executor) {
         executor.transferArtifacts();
         executor.install();
 
-        //Test dist list
-        TestUtils.verifyDistList(executor);
-        //Test installation
-        TestUtils.testInstallation(executor, previousVersion, previousSpecVersion, previousToolVersion);
-
-        //Test `ballerina update`
-        executor.executeCommand("ballerina update", true);
-        TestUtils.testInstallation(executor, previousVersion, previousSpecVersion, toolVersion);
-
-        //Execute all ballerina dist commands once updated
-        TestUtils.testDistCommands(executor, previousVersion, previousSpecVersion, toolVersion, previousVersion,
-                previousSpecVersion, previousVersionsLatestPatch, latestToolVersion);
+        TestUtils.testInstallation(executor, version, specVersion, toolVersion);
 
         executor.uninstall();
         executor.cleanArtifacts();
