+++ test-scripts/ballerina-test-automation/installer-test/src/test/resources/testng.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  ~ Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~ Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
   ~
   ~ WSO2 Inc. licenses this file to you under the Apache License,
   ~ Version 2.0 (the "License"); you may not use this file except
@@ -23,12 +23,9 @@
 <suite name="Suite1" verbose="1">
     <test name="io.ballerina.test">
         <classes>
-            <class name="io.ballerina.test.InstallerTest"/>
-            <class name="io.ballerina.test.UpdateDistTest"/>
-            <class name="io.ballerina.test.UpdateToolTest"/>
-<!--            <class name="io.ballerina.test.CentralTest"/>-->
-            <class name="io.ballerina.test.FetchDependencyTest"/>
-            <class name="io.ballerina.test.ProjectTest"/>
+            <class name="io.ballerina.installer.test.InstallerTest"/>
+            <class name="io.ballerina.installer.test.ProjectTest"/>
+            <class name="io.ballerina.installer.test.CentralTest"/>
         </classes>
     </test>
 </suite>
