+++ test-scripts/ballerina-test-automation/update-tool-test/build.gradle
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+plugins {
+    id 'java'
+}
+
+group 'io.ballerina'
+version '1.0-SNAPSHOT'
+
+sourceCompatibility = 1.8
+
+dependencies {
+    implementation project(':test-automation')
+    testCompile project(path: ':installer-test', configuration: 'testClasses')
+    implementation 'org.testng:testng:6.14.3'
+    testCompile 'org.testng:testng'
+}
+
+repositories {
+    mavenCentral()
+}
+
+test {
+    useTestNG()
+    options.suites('src/test/resources/testng.xml')
+    systemProperty 'BALLERINA_VERSION', project.properties['ballerinaVersion']
+    systemProperty 'SPEC_VERSION', project.properties['specVersion']
+    systemProperty 'TOOL_VERSION', project.properties['toolVersion']
+    systemProperty 'LATEST_TOOL_VERSION', project.properties['latestToolVersion']
+    systemProperty 'LATEST_PATCH_VERSION', project.properties['latestPatchVersion']
+    systemProperty 'LATEST_PATCH_SPEC_VERSION', project.properties['latestPatchSpecVersion']
+}
