+++ test-scripts/ballerina-test-automation/update-tool-test/src/test/java/io/ballerina/tool/test/UpdateToolTest.java
@@ -14,30 +14,49 @@
  * limitations under the License.
  */
 
-package io.ballerina.test;
+package io.ballerina.tool.test;
 
+import io.ballerina.installer.test.TestUtils;
+import io.ballerina.test.Executor;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 
-public class InstallerTest {
+public class UpdateToolTest {
     String version = System.getProperty("BALLERINA_VERSION");
     String specVersion = System.getProperty("SPEC_VERSION");
     String toolVersion = System.getProperty("TOOL_VERSION");
+    String latestToolVersion = System.getProperty("LATEST_TOOL_VERSION");
+
+    String previousVersion = "1.2.0";
+    String previousSpecVersion = "2020R1";
+    String previousVersionsLatestPatch = System.getProperty("LATEST_PATCH_VERSION");
+    String previousToolVersion = "0.8.5";
 
     @DataProvider(name = "getExecutors")
     public Object[][] dataProviderMethod() {
         Executor[][] result = new Executor[1][1];
-        result[0][0] = TestUtils.getExecutor(version);
+        result[0][0] = TestUtils.getExecutor(previousVersion);
         return result;
     }
 
     @Test(dataProvider = "getExecutors")
-    public void testSmoke(Executor executor) {
+    public void testUpdateTool(Executor executor) {
         executor.transferArtifacts();
         executor.install();
 
-        TestUtils.testInstallation(executor, version, specVersion, toolVersion);
+        //Test dist list
+        TestUtils.verifyDistList(executor);
+        //Test installation
+        TestUtils.testInstallation(executor, previousVersion, previousSpecVersion, previousToolVersion);
+
+        //Test `ballerina update`
+        executor.executeCommand("ballerina update", true);
+        TestUtils.testInstallation(executor, previousVersion, previousSpecVersion, toolVersion);
+
+        //Execute all ballerina dist commands once updated
+        TestUtils.testDistCommands(executor, previousVersion, previousSpecVersion, toolVersion, previousVersion,
+                previousSpecVersion, previousVersionsLatestPatch, latestToolVersion);
 
         executor.uninstall();
         executor.cleanArtifacts();
